:pencil2: rename: repository 파일 위치 이동
